backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
alembic==1.14.0
pydantic==2.10.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
minio==7.2.11
httpx==0.28.1
apscheduler==3.10.4
redis==5.2.0
Pillow==11.0.0
qrcode==8.0
reportlab==4.2.5
python-dotenv==1.0.1
aiofiles==24.1.0
email-validator==2.2.0
